Manager Functions As it Should
Fish attaches to cuttingboard correctly and displays the Cut Meter UI correctly and the fish actor destroys itself after the cut. Will have to add functionality to spawn and attach the proper sushi that was cut from the fish. Also figure out why the fish doesnt properly attatch to the player socket